1.在系统管理中，把从数据库中得到的用户信息设置到UI界面中 2.优化数据库代码 3.系统界面中从系统管理的界面到个人信息界面，内部设计模式需要修改 4.完成个人信息界面修改，所有模型和View都在一个文件中

+++ chargePileManage/Classes/Common/Category/NSString+XFTime.m
@@ -14,6 +14,6 @@
     NSInteger min = time / 60;
     NSInteger second = (NSInteger)time % 60;
     
-    return [NSString stringWithFormat:@"%02ld:%02ld", min, second];
+    return [NSString stringWithFormat:@"%02ld:%02ld", (long)min, (long)second];
 }
 @end
